+++ src/app/navigateRequestToComponentRouter.ts
@@ -1,12 +1,12 @@
 import url from 'url';
 import { RequestListener } from 'http';
 
-import { combineResponseWithError } from './combineResponse';
-import { AppError, BadRequestError } from '../constants/error/index';
-import { ROUTE } from '../constants/routes';
-import userRouter from '../routers/userRouter';
+import { combineResponseWithError } from '../shared/server/utils/combineResponse';
+import { AppError, BadRequestError } from '../shared/error/errorInstance';
+import { ROUTE } from '../shared/server/routes';
+import userRouter from '../components/users/user.router';
 
-export const navigateRequestToProcessing: RequestListener = async (request, response) => {
+export const navigateRequestToComponentRouter: RequestListener = async (request, response) => {
   const { pathname } = url.parse(request.url as string);
 
   if (!pathname) {
