+++ src/shared/error/errorInstance/badRequestError.ts
@@ -1,4 +1,4 @@
-import { BAD_REQUEST_ERROR_MESSAGE } from '../error';
+import { BAD_REQUEST_ERROR_MESSAGE } from '../error.message';
 import { AppError } from './appError';
 
 export class BadRequestError extends AppError {
