+++ src/shared/validation/specific/age.validator.ts
@@ -1,0 +1,12 @@
+import { INCORRECT_AGE_ERROR_MESSAGE } from '../../error/error.message';
+import { ValidationError } from '../../error/errorInstance';
+
+export const validateAge = (age: number, minAge: number): void => {
+  const isValidAge = age >= minAge;
+
+  if (!isValidAge) {
+    const errorMessage = INCORRECT_AGE_ERROR_MESSAGE;
+
+    throw new ValidationError(errorMessage);
+  }
+};
