+++ src/shared/validation/validation.constants.ts
@@ -1,0 +1,5 @@
+export enum FIELD_TYPE {
+  ARRAY = 'ARRAY',
+  STRING = 'STRING',
+  NUMBER = 'NUMBER',
+}
